dy-downloader/requirements.txt
aiohttp>=3.9.0
aiofiles>=23.2.1
aiosqlite>=0.19.0
rich>=13.7.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
